Minor fixes

+++ src/Veeam.Signature/Solution/ProgramParameters.cs
@@ -5,8 +5,8 @@ namespace Veeam.Signature.Solution
 {
     public class ProgramParameters
     {
-        [BoundedValueArgument(typeof(int), 'b', "blocksize", MinValue = 1, MaxValue = int.MaxValue, Description = "Block size")]
-        public int BlockSize { get; set; }
+        [BoundedValueArgument(typeof(int), 'b', "blocksize", MinValue = 1, MaxValue = long.MaxValue, Description = "Block size")]
+        public long BlockSize { get; set; }
 
         [FileArgument('i', "input", Description = "Input file", FileMustExist = true)]
         public FileInfo InputFile;
